// src/service/member/schemas/accountV2ForUpdateOrganizationAndPositionClassification.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Member API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 직업 정보
 * @nullable
 */
export type AccountV2ForUpdateOrganizationAndPositionClassification =
  | (typeof AccountV2ForUpdateOrganizationAndPositionClassification)[keyof typeof AccountV2ForUpdateOrganizationAndPositionClassification]
  | null;

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccountV2ForUpdateOrganizationAndPositionClassification = {
  NURSERY_TEACHER: 'NURSERY_TEACHER',
  KINDERGARTEN_TEACHER: 'KINDERGARTEN_TEACHER',
  ELEMENTRAY_SCHOOL_TEACHER: 'ELEMENTRAY_SCHOOL_TEACHER',
  MANAGER: 'MANAGER',
  CONTRIBUTOR: 'CONTRIBUTOR',
  ETC: 'ETC',
} as const;