// src/service/member/schemas/apiResponseAccountResultV2.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Member API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { AccountResultV2 } from './accountResultV2';

export interface ApiResponseAccountResultV2 {
  status: number;
  result?: AccountResultV2;
  timestamp?: string;
  error?: string;
  debug?: string;
}